requirements.txt
absl-py==2.1.0
accelerate==0.31.0
addict==2.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
antiorm==1.2.1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.13.3
bitsandbytes==0.42.0
blinker==1.8.2
boto3==1.36.0
botocore==1.36.0
cachetools==5.5.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contourpy==1.1.0
cycler==0.11.0
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.20.0
db==0.1.1
db-sqlite3==0.0.1
deepface==0.0.93
dill==0.3.8
distlib==0.3.7
Django==5.0.6
django-cors-headers==4.6.0
djangorestframework==3.15.2
docker-pycreds==0.4.0
docstring_parser==0.16
eval_type_backport==0.2.2
fastjsonschema==2.20.0
filelock==3.17.0
fire==0.7.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==25.2.10
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2024.5.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.12.1
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.5.0
itsdangerous==2.2.0
jedi==0.17.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.5
libclang==18.1.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.0.4
multiprocess==0.70.16
namex==0.0.8
narwhals==1.26.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.1
open3d==0.18.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==23.1
pandas==2.1.2
parso==0.7.1
peft==0.11.1
pillow==11.1.0
platformdirs==3.10.0
plotly==5.24.1
pluggy==1.3.0
propcache==0.2.1
protobuf==3.20.3
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.0.9
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jsonrpc-server==0.4.0
pytz==2023.3.post1
PyYAML==6.0.1
razorpay==1.4.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
retina-face==0.0.17
retrying==1.3.4
rich==13.9.4
rpds-py==0.20.1
s3transfer==0.11.0
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.11.4
sentencepiece==0.2.0
sentry-sdk==2.7.1
setproctitle==1.3.3
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sqlparse==0.5.0
streamlit==1.42.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.2.0
tk==0.1.0
together==1.4.0
tokenizers==0.19.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
trl==0.8.6
typer==0.15.1
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2023.3
ujson==5.8.0
unsloth @ git+https://github.com/unslothai/unsloth.git@499635a19013319367358dcf2a0211d30073ebd0
urllib3==2.2.1
virtualenv==20.24.5
wandb==0.17.3
websockets==13.1
Werkzeug==3.0.3
wrapt==1.17.2
xxhash==3.4.1
yarl==1.18.3
zipp==3.20.2
